requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
torch==2.5.1
torchvision==0.21.0
pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==2.0.2
pydantic==2.9.0
python-multipart==0.0.12